+++ docs/release-notes.md
@@ -7,6 +7,10 @@
   * Migrated to Harmony 2.0 (see [_migrate to Harmony 2.0_](https://stardewvalleywiki.com/Modding:Migrate_to_Harmony_2.0) for more info).
 -->
 
+## Upcoming release
+* For players:
+  * Updated for Stardew Valley 1.5.
+
 ## 3.7.6
 Released 21 November 2020 for Stardew Valley 1.4.1 or later.